fixed adding image bug
